Add functions for processing object
