+++ packages/core/src/convertor/wxToWeb.js
@@ -1,8 +1,11 @@
 import * as wxLifecycle from '../platform/patch/wx/lifecycle'
 import * as webLifecycle from '../platform/patch/web/lifecycle'
 import { mergeLifecycle } from './mergeLifecycle'
-import { error } from '../helper/log'
-import { isObject, diffAndCloneA, hasOwn } from '../helper/utils'
+import { diffAndCloneA, hasOwn } from '../helper/utils'
+import {
+  isObject,
+  error
+} from '@mpxjs/utils'
 import { implemented } from '../core/implement'
 import { isDev } from '../helper/env'
