+++ packages/core/src/observer/computed.js
@@ -1,4 +1,5 @@
-import { noop, isFunction } from '../helper/utils'
+import { isFunction } from '../helper/utils'
+import { noop } from '@mpxjs/utils'
 import Dep from './dep'
 import { createRef } from './ref'
 import { ReactiveEffect } from './effect'
