+++ packages/core/src/platform/builtInMixins/renderHelperMixin.js
@@ -1,4 +1,4 @@
-import { isObject } from '../../helper/utils'
+import { isObject } from '@mpxjs/utils'
 
 export default function renderHelperMixin () {
   return {
