+++ packages/pinia/src/util.js
@@ -1,7 +1,5 @@
-import EXPORT_MPX, {
-  isRef,
-  isReactive
-} from '@mpxjs/core'
+import { isRef, isReactive } from '@mpxjs/core'
+import { isPlainObject } from '@mpxjs/utils'
 
 let activePinia
 
@@ -13,30 +11,6 @@ const getActivePinia = () => {
 
 function isComputed (obj) {
   return !!(isRef(obj) && (obj).effect)
-}
-
-function type (n) {
-  return Object.prototype.toString.call(n).slice(8, -1)
-}
-
-function isPlainObject (value) {
-  if (value === null || typeof value !== 'object' || type(value) !== 'Object') return false
-  const proto = Object.getPrototypeOf(value)
-  if (proto === null) return true
-  // 处理支付宝接口返回数据对象的__proto__与js中创建对象的__proto__不一致的问题，判断value.__proto__.__proto__ === null时也认为是plainObject
-  const innerProto = Object.getPrototypeOf(proto)
-  if (proto === Object.prototype || innerProto === null) return true
-  // issue #644
-  if (EXPORT_MPX.config.observeClassInstance) {
-    if (Array.isArray(EXPORT_MPX.config.observeClassInstance)) {
-      for (let i = 0; i < EXPORT_MPX.config.observeClassInstance.length; i++) {
-        if (proto === EXPORT_MPX.config.observeClassInstance[i].prototype) return true
-      }
-    } else {
-      return true
-    }
-  }
-  return false
 }
 
 function mergeReactiveObjects (target, patchToApply) {
