+++ packages/store/src/helper/utils.js
@@ -1,21 +1,5 @@
 import { isRef } from '@mpxjs/core'
-import _getByPath from './getByPath'
-
-const noop = () => {}
-function isObject (obj) {
-  return obj !== null && typeof obj === 'object'
-}
-function isExistAttr (obj, attr) {
-  const type = typeof obj
-  const isNullOrUndefined = obj === null || obj === undefined
-  if (isNullOrUndefined) {
-    return false
-  } else if (type === 'object' || type === 'function') {
-    return attr in obj
-  } else {
-    return obj[attr] !== undefined
-  }
-}
+import { noop, isObject } from '@mpxjs/utils'
 
 export function proxy (target, source, keys, readonly, onConflict) {
   keys = keys || Object.keys(source)
@@ -44,32 +28,6 @@ export function proxy (target, source, keys, readonly, onConflict) {
     Object.defineProperty(target, key, descriptor)
   })
   return target
-}
-
-export function getByPath (data, pathStrOrArr, defaultVal, errTip) {
-  const results = []
-  let normalizedArr = []
-  if (Array.isArray(pathStrOrArr)) {
-    normalizedArr = [pathStrOrArr]
-  } else if (typeof pathStrOrArr === 'string') {
-    normalizedArr = pathStrOrArr.split(',').map(str => str.trim())
-  }
-
-  normalizedArr.forEach(path => {
-    if (!path) return
-    const result = _getByPath(data, path, (value, key) => {
-      let newValue
-      if (isExistAttr(value, key)) {
-        newValue = value[key]
-      } else {
-        newValue = errTip
-      }
-      return newValue
-    })
-    // 小程序setData时不允许undefined数据
-    results.push(result === undefined ? defaultVal : result)
-  })
-  return results.length > 1 ? results : results[0]
 }
 
 export function normalizeMap (prefix, arr) {
