+++ packages/utils/src/common.js
@@ -1,9 +1,35 @@
+import EXPORT_MPX from '@mpxjs/core'
 import _getByPath from './getByPath'
 
 const noop = () => {}
 
 function isObject (obj) {
   return obj !== null && typeof obj === 'object'
+}
+
+function type (n) {
+  return Object.prototype.toString.call(n).slice(8, -1)
+}
+
+function isPlainObject (value) {
+  if (value === null || typeof value !== 'object' || type(value) !== 'Object') return false
+  const proto = Object.getPrototypeOf(value)
+  if (proto === null) return true
+  // 处理支付宝接口返回数据对象的__proto__与js中创建对象的__proto__不一致的问题，判断value.__proto__.__proto__ === null时也认为是plainObject
+  const innerProto = Object.getPrototypeOf(proto)
+  if (proto === Object.prototype || innerProto === null) return true
+  // issue #644
+  const observeClassInstance = EXPORT_MPX.config.observeClassInstance
+  if (observeClassInstance) {
+    if (Array.isArray(observeClassInstance)) {
+      for (let i = 0; i < observeClassInstance.length; i++) {
+        if (proto === observeClassInstance[i].prototype) return true
+      }
+    } else {
+      return true
+    }
+  }
+  return false
 }
 
 function isExistAttr (obj, attr) {
@@ -67,10 +93,11 @@ function normalizeMap (prefix, arr) {
   return arr
 }
 
-
 export {
   noop,
+  type,
   isObject,
+  isPlainObject,
   getByPath,
   normalizeMap
 }
